Sentiment/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
transformers>=4.28.1
torch>=2.0.0
numpy>=1.24.2
motor>=3.1.1
pymongo>=4.3.3
python-dotenv>=1.0.0
httpx>=0.24.0
python-multipart>=0.0.6
huggingface-hub>=0.14.1